debugging metric functions

+++ keras_gpt_2/metrics.py
@@ -93,7 +93,7 @@ def top_3_mc(y_true, y_pred):
     return top_3(y_true, y_pred)
 
 def precision_lm(y_true, y_pred):
-    y_true = tf.cast(y_true, tf.int32)
+    y_true = tf.cast(y_true, tf.int64)
     # y_true = K.reshape(tf.cast(one_hot(y_true, 50257, axis=-1), tf.float32), (-1, 50257))
     y_pred = K.argmax(y_pred, axis=-1)
     return precision_m(y_true, y_pred)
